implemented filter parameter

+++ src/main/java/at/upstream_mobility/itacademy/bored/app/Response.java
@@ -1,6 +1,6 @@
 package at.upstream_mobility.itacademy.bored.app;
 
-public record Response(String activity, String type, int participants, int price,
+public record Response(String activity, String type, int participants, double price,
                        String link, String key, double accessibility) {
 
 }
